added pipe connection functionality and splited the code in different js files
