--- js/map.js
@@ -1,3 +1,0 @@
-(() => {
-    console.info("Map view loaded");
-})();
