+++ map.html
@@ -139,7 +139,8 @@
 	<script type="text/javascript" src="https://revh.github.io/leaflet-draggable-polyline/leaflet.geometryutil.js"></script>
     <script type="text/javascript" src="https://rawgit.com/RevH/leaflet-draggable-polyline/master/leaflet-draggable-polyline.js"></script>
 	<script type="module" src="js/loadLeafletMap.js"></script>
-	<script type="module" src="js/src.js"></script>
+	<script type="module" src="js/add.js"></script>
+	<script type="module" src="js/edit.js"></script>
 	<script type="module" src="js/getKey.js"></script>
 	<script type="text/javascript" src="js/exampleSaveData.json"></script>
    	<script src="https://unpkg.com/esri-leaflet"></script>
